requirements.txt
crewai
requests
python-dotenv
pydantic
